fixes for MacOS X
git-svn-id: https://bugengine.googlecode.com/svn/trunk@361 f3700e90-1ba7-11de-b56b-c9554f25afe1

+++ src/engine/rtti/api/rtti/engine/classinfo.script.hh
@@ -25,8 +25,8 @@ public:
     raw<const ClassInfo>            metaclass;
     mutable raw<const ClassInfo>    child;
     mutable raw<const ClassInfo>    next;
-    size_t                          size;
-    size_t                          propertyCount;
+    u16                             size;
+    u16                             propertyCount;
     raw<PropertyInfo const>         properties;
 private:
     void copy(const void* src, void* dst) const;
